refactor 'setNav0'
